Individual and Corporate customer

+++ src/main/java/growlab/customer/domain/CorporateCustomerShareholder.java
@@ -14,6 +14,6 @@ public class CorporateCustomerShareholder {
     private Integer id;
     private Integer customerId;
     private String shareholder;
-    private Integer sharePercent;
+    private Double sharePercent;
 
 }
